Name the end-of-game specs after the file they test
Each spec now sits next to the module exporting the moves it exercises, so a
game's tests have one obvious home to grow in: distinct-squares splits into a
spec per game, and remove-row-or-column's becomes board-client.spec.ts.
