+++ src/components/games/remove-row-or-column/board-client.spec.ts
